requirements.txt
dantro>=0.20.0b0
h5py>=3.7.0
matplotlib
numpy
pandas
pytest
ruamel.yaml
ruamel.base
scipy
setuptools
torch
utopya>=1.1.3
xarray
coloredlogs
pre-commit
coverage
tqdm
torchdiffeq
